+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ## Unreleased
 
 + Added `JImTexture.fromBytes` for in-memory image loading
-+ Added constructor for `JImGui` which takes a `long` as the native object pointer
-  + This requires you to manually manage the release and initialization of GLFW
++ Added factory for `JImGui` which takes a `long` as the native object pointer
+  + This requires you to manually manage the release and initialization of GLFW/DirectX9
 + Renamed some factories of `JImTexture`
 + Added `windowDrawListAddLine`, `windowDrawListAddImage`
 + Added `inputText` with its flags class